+++ src/ffi/mod.rs
@@ -3,9 +3,7 @@
 mod array;
 mod bridge;
 mod generated;
-#[cfg(feature = "io_ipc")]
-#[cfg_attr(docsrs, doc(cfg(feature = "io_ipc")))]
-pub(crate) mod mmap;
+pub mod mmap;
 mod schema;
 mod stream;
